requirements.txt
pandas>=2.0.0
openpyxl>=3.1.0
edgartools>=0.20.0

















